# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/invitations/{slug}/publish": {
      "post": {
        "description": "Refused until it has an enabled RSVP block, a venue with an address, and a start date in the future. Every missing item is named at once.",
        "operationId": "InvitationLifecycleController_publish",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Make the invitation live",
        "tags": [
          "invitations"
        ]
      }
    },
    "/api/v1/invitations/{slug}/close": {
      "post": {
        "description": "The page stays readable to guests; only new RSVPs are refused.",
        "operationId": "InvitationLifecycleController_close",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Stop accepting responses",
        "tags": [
          "invitations"
        ]
      }
    },
    "/api/v1/invitations/{slug}/reopen": {
      "post": {
        "operationId": "InvitationLifecycleController_reopen",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Accept responses again after closing",
        "tags": [
          "invitations"
        ]
      }
    },
    "/api/v1/invitations/{slug}/send": {
      "post": {
        "description": "Safe to press twice: a guest is only invited once per invitation. Refused while the invitation is a draft, because the link would 404.",
        "operationId": "InvitationLifecycleController_send",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendInvitationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Send the invitation — one email per household",
        "tags": [
          "invitations"
        ]
      }
    },
    "/api/v1/invitations/{slug}/remind": {
      "post": {
        "description": "Only guests who were actually invited and have not replied. At most one reminder per guest per day, so pressing twice is safe.",
        "operationId": "InvitationLifecycleController_remind",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Chase the households that have not answered",
        "tags": [
          "invitations"
        ]
      }
    },
    "/api/v1/invitations/{slug}/notify-changes": {
      "post": {
        "description": "Offered after the date, time or a venue changes; never sent automatically. Only households the invitation reached. Pressing twice in a minute sends once.",
        "operationId": "InvitationLifecycleController_notifyChanges",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotifyChangesDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Tell everyone who holds the invitation that its details changed",
        "tags": [
          "invitations"
        ]
      }
    },
    "/api/v1/invitations/{slug}/thank-you": {
      "post": {
        "description": "Only those who actually checked in, and only once ever. Refused before the event, because a thank-you that arrives first cannot be unsent.",
        "operationId": "InvitationLifecycleController_thankAttendees",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Thank the guests who came",
        "tags": [
          "invitations"
        ]
      }
    },
...
    "/api/v1/events/{eventId}/exports": {
      "get": {
        "operationId": "ExportsController_list",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Past exports, newest first",
        "tags": [
          "exports"
        ]
      },
      "post": {
        "description": "Records the export and returns its downloadPath. The file is built when it is downloaded, by a signed-in caller — there is no public link. A ticket manifest needs guest:contact:read.",
        "operationId": "ExportsController_create",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Ask for an export",
        "tags": [
          "exports"
        ]
      }
    },
    "/api/v1/events/{eventId}/exports/{exportId}": {
      "get": {
        "operationId": "ExportsController_findOne",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "exportId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "One export",
        "tags": [
          "exports"
        ]
      }
    },
    "/api/v1/events/{eventId}/exports/{exportId}/download": {
      "get": {
        "description": "Built now from the event as it is. Email and phone columns appear only for a caller with guest:contact:read.",
        "operationId": "ExportsController_download",
        "parameters": [
          {
            "name": "eventId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "exportId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Download an export",
        "tags": [
          "exports"
        ]
      }
    },
...
  },
  "info": {
    "title": "Aveline API",
    "description": "Event invitations, guest graph, ticketing and operations. See docs/API.md for conventions that apply across every endpoint.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SendInvitationDto": {
        "type": "object",
        "properties": {
          "guestIds": {
            "description": "Send only to these guests’ households. Omitted invites every household on the event that has not been invited yet.",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "NotifyChangesDto": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string",
            "description": "A line from the host, added to the message, e.g. \"We have moved to the garden.\""
          }
        }
      },
...
      "CreateExportDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "GUEST_LIST",
              "SEATING_CHART",
              "PLACE_CARDS",
              "CATERING_SHEET",
              "BAR_SHEET",
              "PLAYLIST",
              "TICKET_MANIFEST"
            ]
          },
          "format": {
            "type": "string",
            "enum": [
              "PDF",
              "CSV",
              "XLSX"
            ],
            "default": "CSV"
          }
        },
        "required": [
          "kind"
        ]
      },
...
    }
  }
}